some conditions added for all operations

+++ src/main/java/com/restaurant/main/menu/MenuService.java
@@ -17,7 +17,8 @@ public class MenuService {
     }
 
     public Menu create(Menu item) {
-
+            if(menuRepo.existsById(item.getId()))
+                return null;
           return  menuRepo.save(item);
     }
 
@@ -26,15 +27,25 @@ public class MenuService {
         return menuRepo.findById(id).orElse(null);
     }
 
-    public void remove(Long id) {
-        menuRepo.deleteById(id);
+    public boolean remove(Long id) {
+        if(menuRepo.existsById(id)) {
+           menuRepo.deleteById(id);
+           return true;
+        }
+        else
+            return false;
+
     }
 
     public Optional<Menu> update(Long id, Menu item) {
-        return menuRepo.findById(id)
-                .map(oldItem -> {
-                    Menu updated = oldItem.updateWith(item);
-                    return menuRepo.save(updated);
-                });
+        if(menuRepo.existsById(id)) {
+            return menuRepo.findById(id)
+                    .map(oldItem -> {
+                        Menu updated = oldItem.updateWith(item);
+                        return menuRepo.save(updated);
+                    });
+
+        }
+        return null;
     }
 }
